requirements.txt
ipython==8.13.2
librosa==0.10.0
matplotlib==3.7.1
numpy==1.24.3
soundfile==0.12.1
echopype==0.7.1
tqdm==4.65.0
pytorch_lightning==1.9.4